Move CIDS from common to core/cids... ... fix everything up after the move. Start work on NTCL. The approach here is to nail down the lowest level routines (i.e. the ones that require very little context to invoke (e.g. wrapping a string in an object that identifies it as, for example, a JSON-encoded onm DAB). The routines that deal with type identification are critical so I'm working on those for the next sprint. Then the wrapper methods on the main core export object will be moved down, and wired into the CRP's bind/dispatch mechanism.
