+++ lib/core/cids/cids.js
