+++ lib/core/crp/crp-bind.js
@@ -39,7 +39,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
 (function() {
   var classRegistry, crp, operationMap;
 
-  classRegistry = require('../../common/onm-class-registry');
+  classRegistry = require('../cids/cids');
 
   operationMap = require('./crp-ops-map');
