+++ lib/core/crp/crp-normalize.js
@@ -39,7 +39,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
 (function() {
   var classRegistry, crp;
 
-  classRegistry = require('../../common/onm-class-registry');
+  classRegistry = require('../cids/cids');
 
   crp = {};
