+++ lib/core/crp/crp-ops-map.js
@@ -39,7 +39,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
 (function() {
   var cids, classRegistry, mapEntry, opId, opTokens, operationDescriptor, operationDescriptorArray, operationMap, _i, _len;
 
-  classRegistry = require('../../common/onm-class-registry');
+  classRegistry = require('../cids/cids');
 
   cids = classRegistry.ids;
