+++ lib/core/onm-core.js
@@ -39,7 +39,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
 (function() {
   var CRP, classRegistry, onm;
 
-  classRegistry = require('../common/onm-class-registry');
+  classRegistry = require('./cids/cids');
 
   CRP = require('./crp/crp');
