+++ lib/core/risp/generate/risp-generate-path-readable.js
@@ -39,7 +39,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
 (function() {
   var classRegistry, xRIP_ReadablePathGenerator;
 
-  classRegistry = require('../../../common/onm-class-registry');
+  classRegistry = require('../../cids/cids');
 
 
   /*
