+++ lib/core/risp/generate/risp-generate.js
@@ -39,7 +39,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
 (function() {
   var classRegistry, xRIP_Generator, xRIP_Generators, xRIP_GeneratorsByFormat;
 
-  classRegistry = require('../../../common/onm-class-registry');
+  classRegistry = require('../../cids/cids');
 
   xRIP_Generators = {
     path: require('./risp-generate-path'),
