+++ lib/core/risp/parse/risp-parse.js
@@ -39,7 +39,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
 (function() {
   var Address, Model, classRegistry, xRIP_Parser, xRIP_parsePath, xRIP_parseVector;
 
-  classRegistry = require('../../../common/onm-class-registry');
+  classRegistry = require('../../cids/cids');
 
   Model = require('../../../onm-model');
