+++ lib/onm-model.js
@@ -38,7 +38,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
   var Address, AddressToken, Model, ModelDetails, classRegistry, helperFunctions, intrinsicDataModels, uuid, xRIP,
     __bind = function(fn, me){ return function(){ return fn.apply(me, arguments); }; };
 
-  classRegistry = require('./common/onm-class-registry');
+  classRegistry = require('./core/cids/cids');
 
   helperFunctions = require('./common/onm-util-functions');
