+++ lib/onm-namespace.js
@@ -40,7 +40,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
   var Address, AddressToken, Namespace, NamespaceDetails, addressResolver, classRegistry,
     __bind = function(fn, me){ return function(){ return fn.apply(me, arguments); }; };
 
-  classRegistry = require('./common/onm-class-registry');
+  classRegistry = require('./core/cids/cids');
 
   AddressToken = require('./core/rasp/onm-address-token');
