+++ lib/onm-store.js
@@ -40,7 +40,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
   var Address, AddressToken, Model, Namespace, Store, StoreDetails, StoreReifier, addressResolver, classRegistry, helperFunctions, uuid,
     __bind = function(fn, me){ return function(){ return fn.apply(me, arguments); }; };
 
-  classRegistry = require('./common/onm-class-registry');
+  classRegistry = require('./core/cids/cids');
 
   helperFunctions = require('./common/onm-util-functions');
