+++ src/core/cids/cids.coffee
