+++ src/core/risp/generate/risp-generate.coffee
@@ -39,7 +39,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
 #
 #
 
-classRegistry = require '../../../common/onm-class-registry'
+classRegistry = require '../../cids/cids'
 
 xRIP_Generators =
     path: require './risp-generate-path'
