+++ src/core/risp/parse/risp-parse.coffee
@@ -39,7 +39,7 @@ BLOG: http://blog.encapsule.org TWITTER: https://twitter.com/Encapsule
 #
 #
 
-classRegistry = require '../../../common/onm-class-registry'
+classRegistry = require '../../cids/cids'
 Model = require '../../../onm-model'
 Address = require '../../../onm-address'
